Fix for problem that test case load-save-load revealed
Problem is when iniparser_dump_ini() function is called, it will
save values without enclosing them with '' or "" marks even if in
original file had them around values.
This will cause that if data has ; or # characters they will be
sanitized away in next load.

Fix is to add "" mark around values that are stored in INI file,
so that when they are read again they are protected.